+++ src/actions.js
@@ -1,30 +1,56 @@
 export default class Actions {
   constructor(tide) {
     this.tide = tide
+    this.namespace = this.namespace || []
   }
 
   getState() {
-    return this.tide.getState()
+    return this.get([])
   }
 
   setState(state, options) {
-    this.tide.setState(state, options)
+    this.setGlobalState(this.getGlobalState().setIn(this.namespace, state), options)
   }
 
   updateState(updater, options) {
-    this.tide.updateState(updater, options)
+    this.setState(updater(this.getState()), options)
   }
 
   mutate(keyPath, value, options) {
-    this.tide.mutate(keyPath, value, options)
+    this.tide.mutate(this._createKeyPath(keyPath), value, options)
   }
 
   get(keyPath) {
-    return this.tide.get(keyPath)
+    return this.tide.get(this._createKeyPath(keyPath))
   }
 
   getActions(name) {
     return this.tide.getActions(name)
+  }
+
+  getGlobalState() {
+    return this.tide.getState()
+  }
+
+  setGlobalState(state, options) {
+    this.tide.setState(state, options)
+  }
+
+  updateGlobalState(updater, options) {
+    this.setGlobalState(updater(this.getGlobalState()), options)
+  }
+
+  mutateGlobal(keyPath, value) {
+    this.tide.mutate(keyPath, value)
+  }
+
+  getGlobal(keyPath) {
+    return this.tide.get(keyPath)
+  }
+
+  _createKeyPath(keyPath) {
+    const kp = Array.isArray(keyPath) ? keyPath : keyPath.split('.')
+    return this.namespace.concat(kp)
   }
 }